+++ lib/panda_doc/api_client.rb
@@ -21,7 +21,11 @@ module PandaDoc
     def initialize(multipart: false)
       @url_prefix = "/public/v1"
       @connection = Faraday.new(PandaDoc.configuration.endpoint) do |conn|
-        conn.authorization :Bearer, PandaDoc.configuration.access_token
+        if PandaDoc.configuration.api_key
+          conn.authorization "API-Key", PandaDoc.configuration.api_key
+        else
+          conn.authorization :Bearer, PandaDoc.configuration.access_token
+        end
 
         if multipart
           conn.request     :multipart